+++ src/main/java/com/xll/service/BasicService.java
@@ -14,7 +14,7 @@ public interface BasicService<T> {
       
     public abstract void delete(int id);  
       
-    public abstract T get(int id);  
+    public abstract T get(long id);  
       
     public abstract List<T> queryAllRecord();
